requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
google-generativeai
python-dotenv
Pillow
plotly
streamlit-pdf-viewer